requirements.txt
asgiref==3.4.1
asttokens==2.0.8
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==5.1.1
Django==3.2.7
django-ckeditor==6.1.0
django-cors-headers==3.13.0
django-filter==21.1
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
executing==1.1.1
gunicorn==20.1.0
idna==3.2
ipython==8.5.0
jedi==0.18.1
Markdown==3.3.4
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow
prompt-toolkit==3.0.31
psycopg2-binary
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.1
traitlets==5.5.0
typing-extensions==4.4.0
urllib3==1.26.7
wcwidth==0.2.5
